Treat the position and depth as a tuple for Stockfish.
When either of these changes, the state transitions and commands sent
to Stockfish are nearly identical.  (The "position" command may be
redundant if only the depth is changed, but the hash table is
preserved anyway, so it's not a problem.)  This means that I can get
instant eval depth changes by just making this change.